TWINS-288: resolving main branch conflicts
