+++ core/src/main/java/org/twins/core/mappers/rest/twinclass/TwinClassRestDTOMapper.java
@@ -149,7 +149,7 @@ public class TwinClassRestDTOMapper extends RestSimpleDTOMapper<TwinClassEntity,
                     .setDeletePermissionId(src.getDeletePermissionId());
         }
         if (mapperContext.hasModeButNot(FeaturerMode.TwinClass2FeaturerMode.HIDE)) {
-            featurerService.loadHeadHunter(src);
+            twinClassService.loadHeadHunter(src);
             dst
                     .setHeadHunterFeaturer(featurerRestDTOMapper.convertOrPostpone(src.getHeadHunterFeaturer(), mapperContext.forkOnPoint(mapperContext.getModeOrUse(FeaturerMode.TwinClass2FeaturerMode.SHORT))))
                     .setHeadHunterFeaturerId(src.getHeadHunterFeaturerId());
@@ -178,7 +178,7 @@ public class TwinClassRestDTOMapper extends RestSimpleDTOMapper<TwinClassEntity,
             twinClassService.loadPermissions(srcCollection);
         }
         if (mapperContext.hasModeButNot(FeaturerMode.TwinClass2FeaturerMode.HIDE)) {
-            featurerService.loadHeadHunter(srcCollection);
+            twinClassService.loadHeadHunter(srcCollection);
         }
     }
